server/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
celery==5.3.6
python-dotenv==1.0.1
pydantic==2.5.3
